% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonic_plant_samples.R
\docType{data}
\name{neonic_plant_samples}
\alias{neonic_plant_samples}
\title{Neonic plant samples}
\format{A data.frame with the following columns:
  
  \code{sample}: the unique plant sample identifier
  
  \code{site}: the site where the sample originated
  
  \code{date}: the date the sample was extracted
  
  \code{plant_species}: species of sample
  
  \code{plant_number}: plant identifier
  
  \code{mass}: mass of sample in grams
  
  \code{acetamiprid}: ng/g acetamiprid detected
  
  \code{clothianidin}: ng/g clothianidin detected
  
  \code{dinotefuran}: ng/g dinotefuran detected
  
  \code{imidacloprid}: ng/g imidacloprid detected
  
  \code{thiacloprid}: ng/g thiacloprid detected
  
  \code{thiamethoxam}: ng/g thiamethoxam detected
  
  \code{surrogate_percent_recovery_imidacloprid_d4}: percent recovery
  imidacloprid}
\usage{
neonic_plant_samples
}
\description{
Contains neonic data from plant samples taken at various STRIPS1 experiment
sites
}
\keyword{datasets}

