% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonic_soil_samples.R
\docType{data}
\name{neonic_soil_samples}
\alias{neonic_soil_samples}
\title{Neonic soil samples}
\format{A data.frame with the following columns:
  \code{sample}: the unique soil sample identifier
  \code{site}: the site where the sample originated
  \code{date}: the date the sample was extracted
  \code{type}: type of sample (see Readme)
  \code{material}: substrate of soil sample
  \code{mass}: mass of sample in grams
  \code{vol}: volume of sample in liters
  \code{organic_carbon_percent}: percent organic carbon
  \code{acetamiprid}: ng/g acetamiprid detected
  \code{clothianidin}: ng/g clothianidin detected
  \code{dinotefuran}: ng/g dinotefuran detected
  \code{imidacloprid}: ng/g imidacloprid detected
  \code{thiacloprid}: ng/g thiacloprid detected
  \code{thiamethoxam}: ng/g thiamethoxam detected
  \code{surrogate_percent_recovery_imidacloprid_d4}: percent recovery imidacloprid}
\usage{
neonic_soil_samples
}
\description{
Contains neonic data from soil samples taken
at various STRIPS1 experiment sites
}
\keyword{datasets}

