% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neonic_water_samples.R
\docType{data}
\name{neonic_water_samples}
\alias{neonic_water_samples}
\title{Neonic water samples}
\format{A data.frame with the following columns:
  \code{sample}: the unique water sample identifier
  \code{site}: the site where the sample originated
  \code{date_start}: the start date of collection
  \code{date_end}: the end date of collection
  \code{time_start}: the start time of collection
  \code{time_end}: the end time of collection
  \code{type}: type of sample (see Readme)
  \code{material}: substrate of water sample
  \code{vol}: volume of sample in liters
  \code{acetamiprid}: ng/g acetamiprid detected
  \code{clothianidin}: ng/g clothianidin detected
  \code{dinotefuran}: ng/g dinotefuran detected
  \code{imidacloprid}: ng/g imidacloprid detected
  \code{thiacloprid}: ng/g thiacloprid detected
  \code{thiamethoxam}: ng/g thiamethoxam detected
  \code{surrogate_percent_recovery_imidacloprid_d4}: percent recovery imidacloprid}
\usage{
neonic_water_samples
}
\description{
Contains neonic data from water samples taken
at various STRIPS1 experiment sites
}
\keyword{datasets}

