% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRIPSHladik.R
\docType{package}
\name{STRIPSHladik}
\alias{STRIPSHladik}
\alias{STRIPSHladik-package}
\title{STRIPSHladik: A package for STRIPS neonic data.}
\description{
The STRIPSHladik package contains STRIPS neonic data
from plant, soil, and water samples
}
\section{STRIPSHladik data sets}{

\code{\link{neonic_plant_samples}}
\code{\link{neonic_soil_samples}}
\code{\link{neonic_water_samples}}
}
\seealso{
\code{\link{neonic_plant_samples}} 
\code{\link{neonic_soil_samples}}
\code{\link{neonic_water_samples}}
}

